requirements.txt
cudatoolkit=11.8
pytorch-gpu
torchvision
tensorboard
opencv
matplotlib
kornia
scikit-image
scikit-learn
scipy
numpy
pycocotools
einops
tqdm
visdom
timm
click
